Moved the DBus interface into a separate module to prevent circular imports.
The logger module requires some constants from the common module, which in turn imported the logger.
This loop is broken by moving the AppService class into a new module.
