+++ lib/autokey/common.py
@@ -17,8 +17,6 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import os
-import dbus.service
-import logging
 
 XDG_CONFIG_HOME = os.environ.get('XDG_CONFIG_HOME', os.path.expanduser('~/.config'))
 # Runtime dir falls back to cache dir, as a fallback is suggested by the spec
@@ -54,28 +52,3 @@ ICON_FILE_NOTIFICATION_DARK = "autokey-status-dark"
 ICON_FILE_NOTIFICATION_ERROR = "autokey-status-error"
 
 USING_QT = False
-
-
-class AppService(dbus.service.Object):
-
-    def __init__(self, app):
-        busName = dbus.service.BusName('org.autokey.Service', bus=dbus.SessionBus())
-        dbus.service.Object.__init__(self, busName, "/AppService")
-        self.app = app
-        logging.debug("Created DBus service")
-
-    @dbus.service.method(dbus_interface='org.autokey.Service', in_signature='', out_signature='')
-    def show_configure(self):
-        self.app.show_configure()
-
-    @dbus.service.method(dbus_interface='org.autokey.Service', in_signature='s', out_signature='')
-    def run_script(self, name):
-        self.app.service.run_script(name)
-
-    @dbus.service.method(dbus_interface='org.autokey.Service', in_signature='s', out_signature='')
-    def run_phrase(self, name):
-        self.app.service.run_phrase(name)
-
-    @dbus.service.method(dbus_interface='org.autokey.Service', in_signature='s', out_signature='')
-    def run_folder(self, name):
-        self.app.service.run_folder(name)
